LayoutTests/imported/w3c/web-platform-tests/webaudio/the-audio-api/the-audioparam-interface/exponentialRamp-special-cases-expected.txt


PASS v0 and v1 have opposite signs
PASS v0 is zero
